+++ src/android/com/outsystems/plugins/healthfitness/OSHealthFitness.kt
@@ -1,7 +1,6 @@
 package com.outsystems.plugins.healthfitness
 
 import android.Manifest
-import android.app.Activity
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Build
@@ -10,9 +9,15 @@ import androidx.core.content.ContextCompat
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
 import com.google.gson.Gson
-import com.outsystems.plugins.healthfitness.store.AdvancedQueryParameters
-import com.outsystems.plugins.healthfitness.store.HealthFitnessManager
-import com.outsystems.plugins.healthfitness.store.HealthStore
+
+import com.outsystems.plugins.healthfitnesslib.background.BackgroundJobParameters
+import com.outsystems.plugins.healthfitnesslib.background.DatabaseManager
+import com.outsystems.plugins.healthfitnesslib.store.AdvancedQueryParameters
+import com.outsystems.plugins.healthfitnesslib.store.HealthFitnessManager
+import com.outsystems.plugins.healthfitnesslib.store.HealthStore
+import com.outsystems.plugins.healthfitnesslib.store.HealthStoreException
+import com.outsystems.plugins.oscordova.CordovaImplementation
+
 import org.apache.cordova.*
 import org.json.JSONArray
 
@@ -25,7 +30,8 @@ class OSHealthFitness : CordovaImplementation() {
     override fun initialize(cordova: CordovaInterface, webView: CordovaWebView) {
         super.initialize(cordova, webView)
         val manager = HealthFitnessManager(cordova.context, cordova.activity)
-        healthStore = HealthStore(this, manager)
+        val database = DatabaseManager(cordova.context)
+        healthStore = HealthStore(cordova.context.applicationContext.packageName, manager, database)
     }
 
     override fun execute(
@@ -35,7 +41,7 @@ class OSHealthFitness : CordovaImplementation() {
     ): Boolean {
         this.callbackContext = callbackContext
 
-        if(!areGooglePlayServicesAvailable(callbackContext)) {
+        if(!areGooglePlayServicesAvailable()) {
             return false;
         }
 
@@ -52,6 +58,9 @@ class OSHealthFitness : CordovaImplementation() {
             "getLastRecord" -> {
                 getLastRecord(args)
             }
+            "setBackgroundJob" -> {
+                setBackgroundJob(args)
+            }
         }
         return true
     }
@@ -66,14 +75,20 @@ class OSHealthFitness : CordovaImplementation() {
         val profileVariables = args.getString(4)
         val summaryVariables = args.getString(5)
 
-        healthStore?.initAndRequestPermissions(
-            customPermissions,
-            allVariables,
-            fitnessVariables,
-            healthVariables,
-            profileVariables,
-            summaryVariables)
-        checkAndGrantPermissions()
+        try {
+            healthStore?.initAndRequestPermissions(
+                customPermissions,
+                allVariables,
+                fitnessVariables,
+                healthVariables,
+                profileVariables,
+                summaryVariables)
+            checkAndGrantPermissions()
+        }
+        catch (hse : HealthStoreException) {
+            sendPluginResult(null, Pair(hse.error.code, hse.error.message))
+        }
+
     }
 
     private fun areAndroidPermissionsGranted(permissions: List<String>): Boolean {
@@ -99,7 +114,9 @@ class OSHealthFitness : CordovaImplementation() {
             if(!healthStore!!.areGoogleFitPermissionsGranted()){
                 setAsActivityResultCallback()
             }
-            healthStore?.requestGoogleFitPermissions()
+            if(healthStore?.requestGoogleFitPermissions() == true) {
+                sendPluginResult("success")
+            }
         }
         else {
             PermissionHelper.requestPermissions(
@@ -112,7 +129,16 @@ class OSHealthFitness : CordovaImplementation() {
 
     private fun advancedQuery(args : JSONArray) {
         val parameters = gson.fromJson(args.getString(0), AdvancedQueryParameters::class.java)
-        healthStore?.advancedQuery(parameters)
+        healthStore?.advancedQueryAsync(
+            parameters,
+            { response ->
+                val pluginResponseJson = gson.toJson(response)
+                sendPluginResult(pluginResponseJson)
+            },
+            { error ->
+                sendPluginResult(null, Pair(error.code, error.message))
+            }
+        )
     }
 
     private fun writeData(args: JSONArray) {
@@ -121,21 +147,60 @@ class OSHealthFitness : CordovaImplementation() {
         val variable = args.getString(0)
         val value = args.getDouble(1).toFloat()
 
-        healthStore?.updateData(variable, value)
+        healthStore?.updateDataAsync(
+            variable,
+            value,
+            { response ->
+                sendPluginResult(response)
+            },
+            { error ->
+                sendPluginResult(null, Pair(error.code, error.message))
+            }
+
+        )
     }
 
     private fun getLastRecord(args: JSONArray) {
         //process parameters
         val variable = args.getString(0)
-        healthStore?.getLastRecord(variable)
+        healthStore?.getLastRecordAsync(
+            variable,
+            { response ->
+                val pluginResponseJson = gson.toJson(response)
+                sendPluginResult(pluginResponseJson)
+            },
+            { error ->
+                sendPluginResult(null, Pair(error.code, error.message))
+            })
+    }
+
+    private fun setBackgroundJob(args: JSONArray) {
+        //process parameters
+        val parameters = gson.fromJson(args.getString(0), BackgroundJobParameters::class.java)
+        healthStore?.setBackgroundJob(
+            parameters,
+            { response ->
+                sendPluginResult(response)
+            },
+            { error ->
+                sendPluginResult(null, Pair(error.code, error.message))
+            }
+        )
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, intent: Intent) {
         //super.onActivityResult(requestCode, resultCode, intent)
-        healthStore?.handleActivityResult(requestCode, resultCode, intent)
+        try {
+            healthStore?.handleActivityResult(requestCode, resultCode, intent)
+        }
+        catch(hse : HealthStoreException) {
+            val error = hse.error
+            sendPluginResult(null, Pair(error.code, error.message))
+        }
+
     }
 
-    override fun areGooglePlayServicesAvailable(callbackContext: CallbackContext): Boolean {
+    override fun areGooglePlayServicesAvailable(): Boolean {
         val googleApiAvailability = GoogleApiAvailability.getInstance()
         val status = googleApiAvailability.isGooglePlayServicesAvailable(cordova.activity)