+++ src/android/com/outsystems/plugins/healthfitness/store/AdvancedQueryResponse.kt
@@ -1,6 +1,0 @@
-package com.outsystems.plugins.healthfitness.store
-
-data class AdvancedQueryResponse(
-    val results : List<AdvancedQueryResponseBlock>,
-    var metadata : String? = ""
-)