+++ src/ios/OSHealthFitness.swift
@@ -1,4 +1,5 @@
 import Foundation
+import OSHealthFitnessLib
 
 @objc(OSHealthFitness)
 class OSHealthFitness: CordovaImplementation {
@@ -65,12 +66,12 @@ class OSHealthFitness: CordovaImplementation {
         let variable = command.arguments[0] as? String ?? ""
         
         plugin?.advancedQuery(variable: variable,
-                        startDate: Date(),
-                        endDate: Date(),
-                        timeUnit: "",
-                        operationType: "",
-                        mostRecent:true,
-                        timeUnitLength: 1) { success, result, error in
+                              startDate: Date.distantPast,
+                              endDate: Date(),
+                              timeUnit: "",
+                              operationType: "MOST_RECENT",
+                              mostRecent:true,
+                              timeUnitLength: 1) { success, result, error in
 
             if error != nil {
                 self.sendResult(result: nil, error: error, callBackID: self.callbackId)
@@ -80,7 +81,7 @@ class OSHealthFitness: CordovaImplementation {
         }
     
     }
-    
+
     @objc(getData:)
     func getData(command: CDVInvokedUrlCommand) {
         callbackId = command.callbackId