+++ src/ios/frameworks/OSCore.xcframework/ios-arm64_x86_64-simulator/OSCore.framework/Headers/OSCore.h
@@ -1,0 +1,17 @@
+//
+//  OSCore.h
+//  OSCore
+//
+//  Created by Carlos Correa on 20/10/2021.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for OSCore.
+FOUNDATION_EXPORT double OSCoreVersionNumber;
+
+//! Project version string for OSCore.
+FOUNDATION_EXPORT const unsigned char OSCoreVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <OSCore/PublicHeader.h>
+